Refactor image domain into model
